Imgur cleanup + added ~rgi command (randomgalleryimage that takes page # and date range)
