+++ src/NadekoBot/Modules/Searches/Searches.cs
@@ -173,95 +173,6 @@ $@"🌍 **Weather for** 【{obj["target"]}】
 
         [NadekoCommand, Usage, Description, Aliases]
         [RequireContext(ContextType.Guild)]
-        public async Task RI(IUserMessage umsg)
-        {
-            var channel = (ITextChannel)umsg.Channel;
-            var provider = GetRandomImgurImage();
-            var link = await provider.ConfigureAwait(false);
-            if (string.IsNullOrWhiteSpace(link))
-                await channel.SendMessageAsync("Search yielded no results ;(").ConfigureAwait(false);
-            else
-                await channel.SendMessageAsync(link).ConfigureAwait(false);
-        }
-
-        [NadekoCommand, Usage, Description, Aliases]
-        [RequireContext(ContextType.Guild)]
-        public async Task Imgur(IUserMessage umsg, [Remainder] string query = null)
-        {
-            var channel = (ITextChannel)umsg.Channel;
-            if (string.IsNullOrWhiteSpace(query))
-            {
-                await channel.SendMessageAsync("Please enter a query ;(").ConfigureAwait(false);
-                return;
-            }
-            var provider = GetImgurImage(query);
-            var link = await provider.ConfigureAwait(false);
-            if (string.IsNullOrWhiteSpace(link))
-                await channel.SendMessageAsync("Search yielded no results ;(").ConfigureAwait(false);
-            else
-                await channel.SendMessageAsync(link).ConfigureAwait(false);
-        }
-
-        public static async Task<string> GetRandomImgurImage()
-        {
-            try
-            {
-                using (var http = new HttpClient())
-                {
-                    http.DefaultRequestHeaders.Clear();
-                    http.DefaultRequestHeaders.Add("Authorization", $"Client-ID {NadekoBot.Credentials.ImgurApiKey}");
-                    var rng = new NadekoRandom();
-                    var reqString = $"https://api.imgur.com/3/gallery/random/random/{rng.Next(1, 50)}";
-                    var obj = JObject.Parse(await http.GetStringAsync(reqString).ConfigureAwait(false));
-                    var items = obj["data"] as JArray;
-                    if (obj["success"].ToString().Equals("false"))
-                    {
-                        return "API call to Imgur was unsuccessful ;(";
-                    }
-                    else
-                    {
-                        return items[rng.Next(1, items.Count)]["link"].ToString();
-                    }
-                }
-            }
-            catch
-            {
-                return "Something went wrong.";
-            }
-        }
-
-        public static async Task<string> GetImgurImage(string query)
-        {
-            if (string.IsNullOrWhiteSpace(query))
-                return "Enter a query!";
-            try
-            {
-                using (var http = new HttpClient())
-                {
-                    http.DefaultRequestHeaders.Clear();
-                    http.DefaultRequestHeaders.Add("Authorization", $"Client-ID {NadekoBot.Credentials.ImgurApiKey}");
-                    var rng = new NadekoRandom();
-                    var reqString = $"https://api.imgur.com/3/gallery/search/top/1?q_any={Uri.EscapeUriString(query)}";
-                    var obj = JObject.Parse(await http.GetStringAsync(reqString).ConfigureAwait(false));
-                    var items = obj["data"] as JArray;
-                    if (obj["success"].ToString().Equals("false"))
-                    {
-                        return "API call to Imgur was unsuccessful ;(";
-                    }
-                    else
-                    {
-                        return items[rng.Next(1, items.Count)]["link"].ToString();
-                    }
-                }
-            }
-            catch
-            {
-                return "Something went wrong.";
-            }
-        }
-
-        [NadekoCommand, Usage, Description, Aliases]
-        [RequireContext(ContextType.Guild)]
         public async Task Lmgtfy(IUserMessage umsg, [Remainder] string ffs = null)
         {
             var channel = (ITextChannel)umsg.Channel;
